WB4BXO: Possibly fixed direct-only digipeating
Since I currently only use aprx as an iGate and fill in digi
(WB4BXO-3) in conjunction with my main digi (WB4BXO-7) I was not
aware of it always only repeating direct heard packets until I
heard about it from kg4pid. I fixed what seemed to me to be the
cause of that problem in digipeater.c where it appears to be
stopping its via search on the first "starred" digi instead of
the first "un-starred" digi.