Refactoring of existing logic in req/done count
Combined every instance of adding the number of requested and done hops
in count_single_tnc2_tracewide() into a single goto slide at the end of
the function.

Replaced the literal comparison with TRACE with the logical question of
(istrace). Perhaps some vestigial code from before the existence of
digipeater.c:match_tracewide()?